containers/services/iot_container/requirements.txt
aiocoap
flask
gunicorn
pika
requests
websockets
paho-mqtt
